# openapi.yaml
openapi: 3.0.0
info:
  title: eBay Monitoring System API
  version: 1.0.0
  description: API for monitoring eBay products, tracking prices, and managing alerts

servers:
  - url: http://localhost:5000/api
    description: Development server

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    type: object

  /auth/login:
    post:
      summary: Login user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '400':
          description: Invalid credentials

  /auth/me:
    get:
      summary: Get current user
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data
        '401':
          description: Unauthorized

  /products:
    get:
      summary: Get all products
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products list
    post:
      summary: Add new product
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ebayUrl]
              properties:
                ebayUrl:
                  type: string
                  format: uri
                supplierUrl:
                  type: string
                  format: uri
      responses:
        '201':
          description: Product created

  /products/{id}:
    get:
      summary: Get product by ID
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product data
    put:
      summary: Update product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted

  /products/{id}/sync:
    post:
      summary: Manually sync product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product synced

  /alerts:
    get:
      summary: Get all alerts
      tags: [Alerts]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Alerts list

  /settings:
    get:
      summary: Get user settings
      tags: [Settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings data
    put:
      summary: Update user settings
      tags: [Settings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                monitoringFrequency:
                  type: integer
                  minimum: 15
                  maximum: 1440
                emailAlerts:
                  type: boolean
      responses:
        '200':
          description: Settings updated

  /metrics:
    get:
      summary: Get dashboard metrics
      tags: [Metrics]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Metrics data

securitySchemes:
  bearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT